move stuff around and add robot commands class
