+++ src/main/java/frc/lib/AllianceFlipUtil.java
@@ -14,7 +14,7 @@ import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.units.measure.Distance;
 import edu.wpi.first.wpilibj.DriverStation;
-import frc.robot.FieldConstants;
+import frc.robot.constants.FieldConstants;
 
 public class AllianceFlipUtil {
   public static Distance applyX(Distance x) {
