+++ src/main/java/frc/robot/RobotCommands.java
@@ -1,0 +1,9 @@
+package frc.robot;
+
+public class RobotCommands {
+    private final RobotState m_state;
+    
+    public RobotCommands(RobotState state) {
+        m_state = state;
+    }
+}
