+++ src/main/java/frc/robot/constants/Constants.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.constants;
 
 import static edu.wpi.first.units.Units.Meters;
 import static edu.wpi.first.units.Units.MetersPerSecond;
